Shutdown traders on unexpected error
Trader will raise errors that are ignored when failing on expected actions (reserving symbols or balance). On all other errors, the graceful shutdown of the ConsumerSupervisor is triggered.

The application supervisor will detect this, and trigger reboot of all of the application's processes. This results in balance being synced properly between exchange and application.