Execute only if 1st trade was updated last
To increase chances of getting to trade, attempt only if the first trade of opportunity was updated last.

Intuition says that if the rest of the trades existed longer, they have a higher chance of being available longer than the last piece to the puzzle.

If multiple bots look at opportunities that become ripe with the same update, the one that takes the latest update will have a chance of successful execution.

Even if a bot starts at the other end, they may execute trades that are not in our opportunity.

This will make Harjus submit order for the key trade multiple milliseconds earlier than before.